+++ src/test/java/org/prgms/TestContextInitializer.java
@@ -1,0 +1,25 @@
+package org.prgms;
+
+import org.springframework.context.ApplicationContextInitializer;
+import org.springframework.context.ConfigurableApplicationContext;
+
+import static com.wix.mysql.EmbeddedMysql.anEmbeddedMysql;
+import static com.wix.mysql.ScriptResolver.classPathScript;
+import static com.wix.mysql.config.Charset.UTF8;
+import static com.wix.mysql.config.MysqldConfig.aMysqldConfig;
+import static com.wix.mysql.distribution.Version.v8_0_11;
+
+public class TestContextInitializer implements ApplicationContextInitializer<ConfigurableApplicationContext> {
+    @Override
+    public void initialize(ConfigurableApplicationContext applicationContext) {
+        var config = aMysqldConfig(v8_0_11)
+                .withCharset(UTF8)
+                .withPort(2215)
+                .withUser("test-user", "test111!")
+                .withTimeZone("Asia/Seoul")
+                .build();
+        anEmbeddedMysql(config)
+                .addSchema("order_mgmt", classPathScript("schema.sql"))
+                .start();
+    }
+}
